+++ src/index.js
@@ -1,11 +1,12 @@
 // import _ from 'lodash';
 // import img from './file.png';
 
-import { Landing } from './landing'
+import { Landing } from './landing';
+import { Contact } from './contact'
 
 // Landing.tabs();
 Landing.header();
 Landing.main();
-Landing.footer();
+Contact.footer();
 
 // console.log(bro('dush'))
